Lift run_detection to core/connection_detect.py
Body byte-identical to the original. Pure stdlib + requests, no
web_server-specific globals or runtime state — no init() needed.

web_server.py: 36500 → 36261 (-239 lines).
